Renamed module "color" to "task-color"
